refactor(session): simplify callbackCount type in concurrent test
Use plain int instead of int32 since the variable is protected by mutex,
not atomic operations. This improves code clarity and consistency.

+++ robot-agent/internal/session/revocation_scenario_test.go
@@ -125,7 +125,7 @@ func TestRevocation_ConcurrentTermination(t *testing.T) {
 	validator := NewTokenValidator(pub, robotID, 30*time.Second)
 	mgr := NewManager(robotID, validator)
 
-	callbackCount := int32(0)
+	callbackCount := 0
 	var mu sync.Mutex
 	mgr.SetStateChangeCallback(func(s State) {
 		if s == StateTerminated {
@@ -157,7 +157,7 @@ func TestRevocation_ConcurrentTermination(t *testing.T) {
 	mu.Lock()
 	count := callbackCount
 	mu.Unlock()
-	assert.Equal(t, int32(1), count)
+	assert.Equal(t, 1, count)
 }
 
 // TestRevocation_ActiveSessionCheck verifies IsActive() behavior.
